Register built-in module types through plugins
