+++ crates/unpack_core/src/asset/mod.rs
@@ -1,2 +1,3 @@
 pub(crate) mod asset_generator;
+pub(crate) mod asset_modules_plugin;
 pub(crate) mod asset_parser;
