+++ crates/unpack_core/src/code_generation.rs
@@ -5,26 +5,24 @@ use std::{
     hash::{Hash, Hasher},
 };
 
-use rspack_sources::{ConcatSource, OriginalSource, RawStringSource, ReplaceSource, SourceMap};
+use rspack_sources::{ConcatSource, RawStringSource, ReplaceSource, SourceMap};
 use serde::{Deserialize, Serialize};
 
 use crate::{
-    AsyncBlockOrigin, AsyncDependenciesBlockIndex, Chunk, ChunkGraph, ChunkGroupKind,
-    CompilerOptions, ConstDependency, Dependency, DependencyIndex, Error, ExportsInfo,
+    AsyncBlockOrigin, AsyncDependenciesBlockIndex, Chunk, ChunkGraph, CompilerOptions,
+    ConstDependency, Dependency, DependencyIndex, Error, ExportsInfo,
     HarmonyExportExpressionDependency, HarmonyExportHeaderDependency,
     HarmonyExportImportedSpecifierDependency, HarmonyExportSpecifierDependency,
     HarmonyImportSideEffectDependency, HarmonyImportSpecifierDependency, ImportDependency, Module,
-    ModuleGraph, ModuleHandle, ModuleType, SourceRange,
+    ModuleGraph, ModuleHandle, SourceRange,
     cache::BuildCache,
     cache_facade::{CacheETag, CacheIdentifier, CacheKey},
     cache_hash::StableHasher,
-    code_generation_record::{
-        CodeGenerationRecord, CodeGenerationReplacement, CodeGenerationResult, CodeGenerationSource,
-    },
+    code_generation_record::{CodeGenerationRecord, CodeGenerationResult, CodeGenerationSource},
     id_assignment::RenderId,
-    output_filename::resolve_chunk_filename,
+    normal_module_factory::{ModuleGeneratorContext, ModuleTypeRegistry},
     rendered_source::RenderedSource,
-    runtime::{RuntimeModule, RuntimeModuleContext, RuntimeRequirement, RuntimeRequirements},
+    runtime::{RuntimeModule, RuntimeRequirement, RuntimeRequirements},
 };
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
@@ -61,17 +59,14 @@ impl CodeGenerationResults {
             .iter()
             .map(|(module, result)| (*module, result.runtime_requirements()))
     }
+
+    pub(crate) fn module_render_id(&self, module: ModuleHandle) -> Option<&RenderId> {
+        self.module_render_ids.get(&module)
+    }
 }
 
-struct CodeGenerationInput<'a> {
-    module: &'a Module,
-    module_graph: &'a ModuleGraph,
-    chunk_graph: &'a ChunkGraph,
-    module_render_ids: &'a HashMap<ModuleHandle, RenderId>,
-}
-
-fn code_generation_etag(input: &CodeGenerationInput<'_>) -> CacheETag {
-    let CodeGenerationInput {
+fn code_generation_etag(input: &ModuleGeneratorContext<'_>) -> CacheETag {
+    let ModuleGeneratorContext {
         module,
         module_graph,
         chunk_graph,
@@ -175,14 +170,28 @@ pub(crate) struct RenderManifest {
     entries: Vec<RenderManifestEntry>,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct RenderManifestEntry {
-    filename: String,
-    render: AssetRenderManifest,
+#[derive(Clone, Copy)]
+pub(crate) struct RenderManifestContext<'a> {
+    pub module_graph: &'a ModuleGraph,
+    pub chunk_graph: &'a ChunkGraph,
+    pub entries: &'a [ModuleHandle],
+    pub code_generation_results: &'a CodeGenerationResults,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
-enum AssetRenderManifest {
+pub(crate) struct RenderManifestEntry {
+    pub filename: String,
+    pub render: RenderManifestContent,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub(crate) enum RenderManifestContent {
+    JavaScript(JavascriptRenderManifest),
+    Asset(Asset),
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub(crate) enum JavascriptRenderManifest {
     InitialChunk {
         modules: Vec<ModuleRenderManifest>,
         runtime_modules: Vec<RenderedRuntimeModule>,
@@ -196,15 +205,15 @@ enum AssetRenderManifest {
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
-struct ModuleRenderManifest {
-    module: ModuleHandle,
-    render_id: RenderId,
+pub(crate) struct ModuleRenderManifest {
+    pub module: ModuleHandle,
+    pub render_id: RenderId,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
-struct RenderedRuntimeModule {
-    module: RuntimeModule,
-    source: String,
+pub(crate) struct RenderedRuntimeModule {
+    pub module: RuntimeModule,
+    pub source: String,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -232,7 +241,7 @@ impl CacheKey for AssetRenderKey {
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
-struct InitFragment {
+pub(crate) struct InitFragment {
     stage: InitFragmentStage,
     order: usize,
     content: String,
@@ -251,13 +260,17 @@ pub(crate) fn generate_code(
     module_graph: &ModuleGraph,
     chunk_graph: &ChunkGraph,
 ) -> CodeGenerationOutcome {
-    generate_code_with(module_graph, chunk_graph, generate_module_code)
+    let module_types = crate::compiler::test_compilation_hooks().normal_module_factory_hooks;
+    generate_code_with(module_graph, chunk_graph, |input| {
+        generate_registered_module(&module_types, input)
+    })
 }
 
 pub(crate) fn generate_code_cached(
     module_graph: &ModuleGraph,
     chunk_graph: &ChunkGraph,
     build_cache: &BuildCache,
+    module_types: &ModuleTypeRegistry,
 ) -> CodeGenerationOutcome {
     let cache = build_cache.code_generations();
     generate_code_with(module_graph, chunk_graph, |input| {
@@ -269,17 +282,29 @@ pub(crate) fn generate_code_cached(
             }
         }
 
-        let record = generate_module_code(input)?;
+        let record = generate_registered_module(module_types, input)?;
         cache.store(key, Some(etag), record.clone());
         Ok(record)
     })
+}
+
+fn generate_registered_module(
+    module_types: &ModuleTypeRegistry,
+    input: ModuleGeneratorContext<'_>,
+) -> Result<CodeGenerationRecord, Error> {
+    if let Some(error) = input.module.build_error() {
+        return Ok(CodeGenerationRecord::new(CodeGenerationSource::Raw {
+            source: render_failed_module_content(error),
+        }));
+    }
+    module_types.generate(input)
 }
 
 fn generate_code_with(
     module_graph: &ModuleGraph,
     chunk_graph: &ChunkGraph,
     mut generate_module: impl FnMut(
-        CodeGenerationInput<'_>,
+        ModuleGeneratorContext<'_>,
     ) -> std::result::Result<CodeGenerationRecord, Error>,
 ) -> CodeGenerationOutcome {
     let module_render_ids = module_graph
@@ -306,7 +331,7 @@ fn generate_code_with(
         .iter()
         .filter(|module| !chunk_graph.module_chunks(module.handle()).is_empty())
     {
-        let input = CodeGenerationInput {
+        let input = ModuleGeneratorContext {
             module,
             module_graph,
             chunk_graph,
@@ -339,76 +364,10 @@ fn generate_code_with(
 }
 
 pub(crate) fn create_render_manifest(
-    chunk_graph: &ChunkGraph,
-    entries: &[ModuleHandle],
-    code_generation_results: &CodeGenerationResults,
+    context: RenderManifestContext<'_>,
+    render_manifest: &crate::compilation::RenderManifestHook,
 ) -> RenderManifest {
-    let mut manifest_entries = Vec::new();
-
-    for (entry_index, group_handle) in chunk_graph.entrypoints().iter().copied().enumerate() {
-        let group = &chunk_graph.chunk_groups()[group_handle.index()];
-        let chunk_handle = group
-            .chunks()
-            .first()
-            .copied()
-            .expect("Entrypoint must contain a Chunk before manifest creation");
-        let chunk = chunk_graph
-            .chunk(chunk_handle)
-            .expect("Entrypoint Chunk must exist before manifest creation");
-        let entry_module = entries
-            .get(entry_index)
-            .copied()
-            .expect("Entrypoint must have an Entry Module before manifest creation");
-        let modules = module_render_manifest(chunk_graph, chunk, code_generation_results);
-        chunk_graph
-            .runtime_tree_requirements(group_handle)
-            .expect("Runtime Requirements must be processed before manifest creation");
-        let runtime_context = RuntimeModuleContext {
-            chunk_graph,
-            runtime_chunk: chunk_handle,
-        };
-        let runtime_modules = chunk_graph
-            .runtime_modules(chunk_handle)
-            .iter()
-            .map(|module| RenderedRuntimeModule {
-                module: *module,
-                source: module.generate(&runtime_context),
-            })
-            .collect();
-        manifest_entries.push(RenderManifestEntry {
-            filename: resolve_chunk_filename(chunk),
-            render: AssetRenderManifest::InitialChunk {
-                modules,
-                runtime_modules,
-                entry_id: code_generation_results
-                    .module_render_ids
-                    .get(&entry_module)
-                    .expect("Entry Module must have a Render ID before manifest creation")
-                    .clone(),
-                chunk_id: chunk.render_id().clone(),
-            },
-        });
-    }
-
-    for chunk in chunk_graph.chunks() {
-        let is_initial = chunk.groups().iter().any(|group_handle| {
-            matches!(
-                chunk_graph.chunk_groups()[group_handle.index()].kind(),
-                ChunkGroupKind::Entrypoint { .. }
-            )
-        });
-        if is_initial {
-            continue;
-        }
-        manifest_entries.push(RenderManifestEntry {
-            filename: resolve_chunk_filename(chunk),
-            render: AssetRenderManifest::AsyncChunk {
-                modules: module_render_manifest(chunk_graph, chunk, code_generation_results),
-                chunk_id: chunk.render_id().clone(),
-            },
-        });
-    }
-
+    let mut manifest_entries = render_manifest.call(context);
     manifest_entries.sort_by(|left, right| left.filename.cmp(&right.filename));
     RenderManifest {
         entries: manifest_entries,
@@ -418,8 +377,6 @@ pub(crate) fn create_render_manifest(
 pub(crate) fn render_assets(
     options: &CompilerOptions,
     build_cache: &BuildCache,
-    module_graph: &ModuleGraph,
-    chunk_graph: &ChunkGraph,
     manifest: &RenderManifest,
     code_generation_results: &CodeGenerationResults,
 ) -> Vec<Asset> {
@@ -427,18 +384,25 @@ pub(crate) fn render_assets(
     let cache = build_cache.asset_renders::<AssetRenderKey>();
     let cache_enabled = options.cache.kind == crate::CacheKind::Filesystem;
     for entry in &manifest.entries {
-        let key = entry.render.cache_key();
+        let RenderManifestContent::JavaScript(render) = &entry.render else {
+            let RenderManifestContent::Asset(asset) = &entry.render else {
+                unreachable!()
+            };
+            assets.push(asset.clone());
+            continue;
+        };
+        let key = render.cache_key();
         let rendered_source = if cache_enabled {
-            let etag = entry.render.cache_etag(code_generation_results);
+            let etag = render.cache_etag(code_generation_results);
             if let Some(rendered_source) = cache.get(&key, Some(&etag)) {
                 rendered_source.as_ref().clone()
             } else {
-                let rendered_source = render_asset(&entry.render, code_generation_results);
+                let rendered_source = render_asset(render, code_generation_results);
                 cache.store(key, Some(etag), rendered_source.clone());
                 rendered_source
             }
         } else {
-            render_asset(&entry.render, code_generation_results)
+            render_asset(render, code_generation_results)
         };
         assets.extend(emit_asset(
             entry.filename.clone(),
@@ -446,14 +410,10 @@ pub(crate) fn render_assets(
             options.sourcemap,
         ));
     }
-    assets.extend(crate::asset::asset_generator::render_resource_assets(
-        module_graph,
-        chunk_graph,
-    ));
     assets
 }
 
-impl AssetRenderManifest {
+impl JavascriptRenderManifest {
     fn cache_key(&self) -> AssetRenderKey {
         match self {
             Self::InitialChunk { chunk_id, .. } => AssetRenderKey {
@@ -516,7 +476,7 @@ fn hash_module_render_inputs(
     }
 }
 
-fn module_render_manifest(
+pub(crate) fn module_render_manifest(
     chunk_graph: &ChunkGraph,
     chunk: &Chunk,
     code_generation_results: &CodeGenerationResults,
@@ -546,17 +506,17 @@ fn module_render_manifest(
 }
 
 fn render_asset(
-    manifest: &AssetRenderManifest,
+    manifest: &JavascriptRenderManifest,
     code_generation_results: &CodeGenerationResults,
 ) -> RenderedSource {
     let source = match manifest {
-        AssetRenderManifest::InitialChunk {
+        JavascriptRenderManifest::InitialChunk {
             modules,
             runtime_modules,
             entry_id,
             chunk_id: _,
         } => render_initial_asset(modules, runtime_modules, entry_id, code_generation_results),
-        AssetRenderManifest::AsyncChunk { modules, chunk_id } => {
+        JavascriptRenderManifest::AsyncChunk { modules, chunk_id } => {
             render_async_chunk_asset(modules, chunk_id, code_generation_results)
         }
     };
@@ -695,112 +655,11 @@ fn render_module_table(
     source
 }
 
-fn generate_module_code(
-    input: CodeGenerationInput<'_>,
-) -> std::result::Result<CodeGenerationRecord, Error> {
-    let CodeGenerationInput {
-        module,
-        module_graph,
-        chunk_graph,
-        module_render_ids,
-    } = input;
-    if let Some(error) = module.build_error() {
-        return Ok(CodeGenerationRecord::new(CodeGenerationSource::Raw {
-            source: render_failed_module_content(error),
-        }));
-    }
-
-    if module.identity().module_type == ModuleType::Json {
-        return Ok(crate::json::json_generator::generate(module));
-    }
-    if module.identity().module_type.is_asset() {
-        return Ok(crate::asset::asset_generator::generate(module));
-    }
-
-    let module_handle = module.handle();
-    let module_render_id = &module_render_ids[&module_handle];
-    let module_render_name = module_render_id.to_string();
-    let mut source = ReplaceSource::new(OriginalSource::new(
-        module.source(),
-        module_render_name.as_str(),
-    ));
-    let mut init_fragments = Vec::new();
-    let mut runtime_requirements = RuntimeRequirements::default();
-    if module.is_harmony() {
-        apply_harmony_compatibility_template(&mut runtime_requirements, &mut init_fragments);
-    }
-
-    for dependency in module.presentational_dependencies() {
-        apply_dependency_template(
-            dependency,
-            module_handle,
-            None,
-            None,
-            module_graph,
-            chunk_graph,
-            module.exports_info(),
-            module_render_ids,
-            &mut runtime_requirements,
-            &mut source,
-            &mut init_fragments,
-        )?;
-    }
-    for (dependency_index, dependency) in module.dependencies().iter().enumerate() {
-        apply_dependency_template(
-            dependency,
-            module_handle,
-            None,
-            Some(DependencyIndex::new(dependency_index)),
-            module_graph,
-            chunk_graph,
-            module.exports_info(),
-            module_render_ids,
-            &mut runtime_requirements,
-            &mut source,
-            &mut init_fragments,
-        )?;
-    }
-    for (block_index, block) in module.blocks().iter().enumerate() {
-        for (dependency_index, dependency) in block.dependencies().iter().enumerate() {
-            apply_dependency_template(
-                dependency,
-                module_handle,
-                Some(AsyncDependenciesBlockIndex::new(block_index)),
-                Some(DependencyIndex::new(dependency_index)),
-                module_graph,
-                chunk_graph,
-                module.exports_info(),
-                module_render_ids,
-                &mut runtime_requirements,
-                &mut source,
-                &mut init_fragments,
-            )?;
-        }
-    }
-
-    let init = render_init_fragments(init_fragments);
-    Ok(
-        CodeGenerationRecord::new(CodeGenerationSource::OriginalWithReplacements {
-            prefix: init,
-            original_source_len: u32::try_from(module.source_len())
-                .expect("Module source length must fit the Code Generation cache format"),
-            original_name: module_render_name,
-            replacements: source
-                .replacements()
-                .iter()
-                .map(CodeGenerationReplacement::from)
-                .collect(),
-            suffix: String::new(),
-        })
-        .with_runtime_requirements(runtime_requirements),
-    )
-}
-
 fn render_failed_module_content(error: &Error) -> String {
     format!("throw new Error({});", json_string(&error.to_string()))
 }
 
-fn apply_harmony_compatibility_template(
+pub(crate) fn apply_harmony_compatibility_template(
     runtime_requirements: &mut RuntimeRequirements,
     init_fragments: &mut Vec<InitFragment>,
 ) {
@@ -812,7 +671,7 @@ fn apply_harmony_compatibility_template(
     );
 }
 
-fn render_init_fragments(mut fragments: Vec<InitFragment>) -> String {
+pub(crate) fn render_init_fragments(mut fragments: Vec<InitFragment>) -> String {
     fragments.sort_by_key(|fragment| (fragment.stage, fragment.order));
     fragments
         .into_iter()
@@ -833,7 +692,7 @@ fn push_init_fragment(
 }
 
 #[allow(clippy::too_many_arguments)]
-fn apply_dependency_template(
+pub(crate) fn apply_dependency_template(
     dependency: &Dependency,
     module_handle: ModuleHandle,
     origin_block: Option<AsyncDependenciesBlockIndex>,
@@ -1223,9 +1082,9 @@ mod tests {
     };
 
     use super::{
-        AssetRenderKey, AssetRenderKind, AssetRenderManifest, CodeGenerationResult,
-        CodeGenerationResults, CodeGenerationSource, ModuleRenderManifest, RenderedRuntimeModule,
-        RenderedSource, RuntimeRequirement, emit_asset,
+        AssetRenderKey, AssetRenderKind, CodeGenerationResult, CodeGenerationResults,
+        CodeGenerationSource, JavascriptRenderManifest, ModuleRenderManifest,
+        RenderedRuntimeModule, RenderedSource, RuntimeRequirement, emit_asset,
     };
     use crate::code_generation_record::{CodeGenerationRecord, CodeGenerationReplacement};
 
@@ -1260,7 +1119,7 @@ mod tests {
             module: module_handle,
             render_id: RenderId::String("./src/feature.js".to_string()),
         };
-        let render = AssetRenderManifest::AsyncChunk {
+        let render = JavascriptRenderManifest::AsyncChunk {
             modules: vec![module],
             chunk_id: RenderId::String("src_feature_js".to_string()),
         };
@@ -1287,7 +1146,7 @@ mod tests {
         );
         assert_ne!(without_requirement, render.cache_etag(&results));
 
-        let initial = AssetRenderManifest::InitialChunk {
+        let initial = JavascriptRenderManifest::InitialChunk {
             modules: Vec::new(),
             runtime_modules: vec![RenderedRuntimeModule {
                 module: RuntimeModule::GetChunkFilename,
@@ -1296,7 +1155,7 @@ mod tests {
             entry_id: RenderId::String("./src/index.js".to_string()),
             chunk_id: RenderId::String("main".to_string()),
         };
-        let changed_filename_map = AssetRenderManifest::InitialChunk {
+        let changed_filename_map = JavascriptRenderManifest::InitialChunk {
             modules: Vec::new(),
             runtime_modules: vec![RenderedRuntimeModule {
                 module: RuntimeModule::GetChunkFilename,
@@ -1345,7 +1204,9 @@ mod tests {
         let build_cache = BuildCache::new(CacheOptions::memory(), SnapshotOptions::default());
 
         let (first_graph, first_chunks, first_module) = build("first");
-        let first = super::generate_code_cached(&first_graph, &first_chunks, &build_cache);
+        let module_types = crate::compiler::test_compilation_hooks().normal_module_factory_hooks;
+        let first =
+            super::generate_code_cached(&first_graph, &first_chunks, &build_cache, &module_types);
         assert_eq!(
             first.results.results[&first_module]
                 .source()
@@ -1355,7 +1216,8 @@ mod tests {
         );
 
         let (second_graph, second_chunks, second_module) = build("second");
-        let second = super::generate_code_cached(&second_graph, &second_chunks, &build_cache);
+        let second =
+            super::generate_code_cached(&second_graph, &second_chunks, &build_cache, &module_types);
         assert_eq!(
             second.results.results[&second_module]
                 .source()
@@ -1400,7 +1262,7 @@ mod tests {
         let module_ref = module_graph
             .module(module)
             .expect("fixture Module should exist");
-        let input = super::CodeGenerationInput {
+        let input = crate::normal_module_factory::ModuleGeneratorContext {
             module: module_ref,
             module_graph: &module_graph,
             chunk_graph: &chunk_graph,
@@ -1427,7 +1289,12 @@ mod tests {
             }),
         );
 
-        let outcome = super::generate_code_cached(&module_graph, &chunk_graph, &build_cache);
+        let outcome = super::generate_code_cached(
+            &module_graph,
+            &chunk_graph,
+            &build_cache,
+            &crate::compiler::test_compilation_hooks().normal_module_factory_hooks,
+        );
         assert_eq!(
             outcome.results.results[&module]
                 .source()
@@ -1468,8 +1335,18 @@ mod tests {
 
         let build_cache = BuildCache::new(CacheOptions::memory(), SnapshotOptions::default());
         for outcome in [
-            super::generate_code_cached(&module_graph, &chunk_graph, &build_cache),
-            super::generate_code_cached(&module_graph, &chunk_graph, &build_cache),
+            super::generate_code_cached(
+                &module_graph,
+                &chunk_graph,
+                &build_cache,
+                &crate::compiler::test_compilation_hooks().normal_module_factory_hooks,
+            ),
+            super::generate_code_cached(
+                &module_graph,
+                &chunk_graph,
+                &build_cache,
+                &crate::compiler::test_compilation_hooks().normal_module_factory_hooks,
+            ),
         ] {
             assert_eq!(
                 outcome.errors,
@@ -1564,7 +1441,7 @@ mod tests {
             compilation.chunk_graph(),
             |input| {
                 generation_count += 1;
-                super::generate_module_code(input)
+                crate::javascript::javascript_generator::generate(input)
             },
         );
         assert!(outcome.errors.is_empty());
@@ -1617,7 +1494,7 @@ mod tests {
                         message: "fixture generation failure".to_string(),
                     })
                 } else {
-                    super::generate_module_code(input)
+                    crate::javascript::javascript_generator::generate(input)
                 }
             },
         );
