+++ crates/unpack_core/src/compilation.rs
@@ -17,7 +17,7 @@ use crate::{
 use tracing::Instrument;
 
 mod hooks;
-pub(crate) use hooks::CompilationHookSet;
+pub(crate) use hooks::{CompilationHookSet, RenderManifestHook};
 
 #[derive(Debug, Clone)]
 pub struct Compilation {
@@ -125,6 +125,7 @@ impl Compilation {
                 self.resolver.clone(),
                 self.build_cache.clone(),
                 self.file_system_info.clone(),
+                self.hooks.normal_module_factory_hooks.clone(),
                 Arc::clone(&state),
             )
             .await;
@@ -221,12 +222,18 @@ impl Compilation {
     }
 
     pub fn create_asset_render_manifest(&mut self) {
+        let code_generation_results = self
+            .code_generation_results
+            .as_ref()
+            .expect("code generation results should exist before render manifest creation");
         self.asset_render_manifest = Some(code_generation::create_render_manifest(
-            &self.chunk_graph,
-            &self.entries,
-            self.code_generation_results
-                .as_ref()
-                .expect("code generation results should exist before render manifest creation"),
+            code_generation::RenderManifestContext {
+                module_graph: &self.module_graph,
+                chunk_graph: &self.chunk_graph,
+                entries: &self.entries,
+                code_generation_results,
+            },
+            &self.hooks.render_manifest,
         ));
     }
 
@@ -234,8 +241,6 @@ impl Compilation {
         self.assets = code_generation::render_assets(
             &self.options,
             &self.build_cache,
-            &self.module_graph,
-            &self.chunk_graph,
             self.asset_render_manifest
                 .as_ref()
                 .expect("render manifest should exist before Asset rendering"),
@@ -256,6 +261,7 @@ impl Compilation {
             &self.module_graph,
             &self.chunk_graph,
             &self.build_cache,
+            &self.hooks.normal_module_factory_hooks,
         );
         self.errors.extend(outcome.errors);
         self.code_generation_results = Some(outcome.results);
