+++ crates/unpack_core/src/javascript/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod javascript_generator;
+pub(crate) mod javascript_modules_plugin;
+pub(crate) mod javascript_parser;
