+++ crates/unpack_core/src/json/mod.rs
@@ -1,2 +1,3 @@
 pub(crate) mod json_generator;
+pub(crate) mod json_modules_plugin;
 pub(crate) mod json_parser;
